# rework Version handling of commit versions