--- a/lib/spack/spack/version.py
+++ b/lib/spack/spack/version.py
@@ -29,10 +29,10 @@ import os
 import re
 from bisect import bisect_left
 from functools import wraps
+from typing import List, Optional, Sequence, Tuple, Union
 
 from six import string_types
 
-import llnl.util.tty as tty
 from llnl.util.filesystem import mkdirp, working_dir
 
 import spack.caches
@@ -52,9 +52,10 @@ COMMIT_VERSION = re.compile(r'^[a-f0-9]{40}$')
 
 # regex for version segments
 SEGMENT_REGEX = re.compile(r'(?:(?P<num>[0-9]+)|(?P<str>[a-zA-Z]+))(?P<sep>[_.-]*)')
+SEPARATORS_T = Tuple[Tuple[Optional[str], Optional[str], str], ...]
 
 # regular expression for semantic versioning
-SEMVER_REGEX = re.compile(".+(?P<semver>([0-9]+)[.]([0-9]+)[.]([0-9]+)"
+SEMVER_REGEX = re.compile(".*?(?P<semver>([0-9]+)[.]([0-9]+)[.]([0-9]+)"
                           "(?:-([0-9A-Za-z-]+(?:[.][0-9A-Za-z-]+)*))?"
                           "(?:[+][0-9A-Za-z-]+)?)")
 
@@ -70,6 +71,9 @@ infinity_versions = [
 ]
 
 iv_min_len = min(len(s) for s in infinity_versions)
+
+VERSION_ISH_T = Union[int, str, 'Version']
+VERSION_TYPES = Union['Version', 'VersionRange', 'VersionList']
 
 
 def coerce_versions(a, b):
@@ -134,6 +138,9 @@ class VersionStrComponent(object):
     def __str__(self):
         return self.data
 
+    def __repr__(self):
+        return "VersionStrComponent({})".format(self.data)
+
     def __eq__(self, other):
         if isinstance(other, VersionStrComponent):
             return self.data == other.data
@@ -180,10 +187,11 @@ class Version(object):
         "separators",
         "string",
         "is_commit",
+        "commit_lookup",
     ]
 
     def __init__(self, string):
-        # type: (str) -> None
+        # type: (VERSION_ISH_T) -> None
         if not isinstance(string, str):
             string = str(string)
 
@@ -194,14 +202,19 @@ class Version(object):
         if string and not VALID_VERSION.match(string):
             raise ValueError("Bad characters in version string: %s" % string)
 
-        # An object that can lookup git commits to compare them to versions
         segments = SEGMENT_REGEX.findall(string)
-        self.version = tuple(
-            int(m[0]) if m[0] else VersionStrComponent(m[1]) for m in segments
-        )
-        self.separators = tuple(m[2] for m in segments)
 
+        self.commit_lookup = None
         self.is_commit = len(self.string) == 40 and COMMIT_VERSION.match(self.string)
+        self.version = tuple()  # type: Tuple[Union[int,VersionStrComponent], ...]
+        self.separators = tuple()  # type: SEPARATORS_T
+        if self.is_commit:
+            self.version = (VersionStrComponent("unknown_git_commit"),)
+        else:
+            self.version = tuple(
+                int(m[0]) if m[0] else VersionStrComponent(m[1]) for m in segments
+            )
+            self.separators = tuple(m[2] for m in segments)
 
     def _cmp(self, commit_lookup):
         if self.is_commit and commit_lookup:
@@ -483,11 +496,12 @@ class Version(object):
         """
 
         # Sanity check we have a commit
-        if not self.is_commit:
-            tty.die("%s is not a commit." % self)
+        if (not self.is_commit) or self.commit_lookup is not None:
+            return
 
         # Generate a commit looker-upper
         cl = CommitLookup(pkg_name)
+        self.commit_lookup = cl
         self.version = self._cmp(cl)
 
 
@@ -512,6 +526,14 @@ class VersionRange(object):
         min_len = min(len(start), len(end))
         if end.up_to(min_len) < start.up_to(min_len):
             raise ValueError("Invalid Version range: %s" % self)
+
+    @property
+    def is_commit(self):
+        if self.start is not None and self.start.is_commit:
+            return True
+        if self.end is not None and self.end.is_commit:
+            return True
+        return False
 
     def lowest(self):
         return self.start
@@ -689,22 +711,37 @@ class VersionRange(object):
             out += str(self.end)
         return out
 
+    def generate_commit_lookup(self, pkg_name):
+        if self.start is not None and self.start.is_commit:
+            self.start.generate_commit_lookup(pkg_name)
+        if self.end is not None and self.end.is_commit:
+            self.end.generate_commit_lookup(pkg_name)
+
 
 class VersionList(object):
     """Sorted, non-redundant list of Versions and VersionRanges."""
 
     def __init__(self, vlist=None):
-        self.versions = []
+        # type: (Optional[Union[Sequence[VERSION_ISH_T],str,VersionList]]) -> None
+        self.versions = []  # type: List[Union[Version, VersionRange]]
         if vlist is not None:
             if isinstance(vlist, string_types):
-                vlist = _string_to_version(vlist)
-                if type(vlist) == VersionList:
-                    self.versions = vlist.versions
+                version_list = _string_to_version(vlist)
+                if isinstance(version_list, VersionList):
+                    self.versions = version_list.versions
                 else:
-                    self.versions = [vlist]
+                    self.versions = [version_list]
             else:
                 for v in vlist:
                     self.add(ver(v))
+
+    @property
+    def is_commit(self):
+        # type: () -> bool
+        for v in self.versions:
+            if v.is_commit:
+                return True
+        return False
 
     def add(self, version):
         if type(version) in (Version, VersionRange):
@@ -712,7 +749,7 @@ class VersionList(object):
             if version.concrete:
                 version = version.concrete
 
-            i = bisect_left(self, version)
+            i = bisect_left(self.versions, version)
 
             while i - 1 >= 0 and version.overlaps(self[i - 1]):
                 version = version.union(self[i - 1])
@@ -873,11 +910,11 @@ class VersionList(object):
             return False
 
         for version in other:
-            i = bisect_left(self, other)
+            i = bisect_left(self.versions, other)
             if i == 0:
                 if version not in self[0]:
                     return False
-            elif all(version not in v for v in self[i - 1:]):
+            elif all(version not in v for v in self.versions[i - 1:]):
                 return False
 
         return True
@@ -930,8 +967,13 @@ class VersionList(object):
     def __repr__(self):
         return str(self.versions)
 
+    def generate_commit_lookup(self, pkg_name):
+        for v in self.versions:
+            v.generate_commit_lookup(pkg_name)
+
 
 def _string_to_version(string):
+    # type: (str) -> VERSION_TYPES
     """Converts a string to a Version, VersionList, or VersionRange.
        This is private.  Client code should use ver().
     """