This makes two main changes:

* Commit versions that have not been looked up, and those for which the
  lookup failed, are treated as having the version "unknown_git_commit"
  which is a new infinity version so they build with the newest build
  instructions in the package.
* Commit version lookups occur on call to `v.generate_commit_lookup()`
  rather than delaying until later.  This means both that it's eagerly
  evaluated and that there is no opportunity for the other side to
  provide a lookup object.  The latter may cause problems if the version
  object is replaced somewhere without having its package attached, but
  it never happens in the tests, or in any build I've tried with this
  branch.

There are a couple of additional changes to fix lingering git related
test issues here that are not directly related but were required to get
the test board green on my end.

It would be really good to get some testing with this on something like MARBL that makes heavy use of commit versions.