@trws 2022-05-06
Some version comparison benchmark numbers:

Before:
```python-console
In [5]: %timeit v1 < v2
968 ns ± 13.1 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

After:
```python-console
In [6]: %timeit v1 < v2
381 ns ± 2.48 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

So about 2.5x.