@becker33 2022-06-21
The eager evaluation scares me for versions in spec matrices. Have you tried

```
spack:
  specs:
  - matrix:
    - [foo]
    - [@<commit1>, @<commit>]
    exclude:
      - '@:9999'
```
Where one commit is develop-like and the other is between numeric versions?