@trws 2022-06-21
I had not @becker33, testing this matrix (first commit is current raja develop second is just before last release):

```yaml
spack:
  specs:
  - matrix:
    - [raja]
    - ['@30e0d44bdf639f1cd8461373442139cdc7e6c640', '@3e863ab64987260d356cd223b68e0075f73a49b1']
    exclude:
      - '@:9999'
```

Before the change, the matrix would install both, after the change it rejected both originally because the versions didn't have package info. That was happening because specs were comparing versions without providing them with `generate_commit_lookup()`, which seems like a bug.  I have a patch to fix that.  Also fixing the fun bug in our semver regex that makes it ignore all but one leading version number in the major version, so raja's v2022.03.0 was parsing as 2.3.0. 😬

After fixing those bugs, nothing is installed for the above because the versions are correctly inferred and neither is greater than 9999.  If I set it to 2000 then it installs the develop version from after 2022.03.0, and not the one from before.  Cleaning up the commit some, will push shortly.