@trws 2022-06-23
It's possible we do, because the code was never set up to set package information on versions correctly in the first place.  The spec class clearly needed some more work in that regard for example.  I'm relatively sure we had cases where it wasn't working before either, and just never noticed.  That's a danger we have from the fact that unknown git versions still get treated as valid, with GitVersion or just as easily with VersionStrComponent about the best we can do is say "if self or other is a correctly resolved commit and target is an unresolved commit, it might possibly be an error."  I suppose we could make that case do a cross-lookup even, the comparisons between those are rare enough that the cost isn't huge especially if we do it by method replacement on the un-resolved object.

To be honest, I'm not all that enthused about rebasing this on a PR that's going to be extremely high conflict, especially given that the PR in question is far, far after this one (admittedly that's probably partly frustration with dealing with the OpenMP committee this week ><).   If that's what needs to happen I suppose I will.