@scheibelp 2022-06-23
> Commit versions that have not been looked up, and those for which the lookup failed, are treated as having the version "unknown_git_commit" which is a new infinity version so they build with the newest build instructions in the package.

I'm concerned with the intent of this PR: when would it be valid to compare an unresolved git version?