@trws 2022-06-23
That has always been the default, except that it used to be essentially `Version("''.0")` or some similar version less than zero.  The intent of this PR is to make it more clear when a version will actually be resolved and when not, and improve the performance of dealing with a mix of regular and commit versions, and change the default to treat an unknown version as higher than any known version on the theory it is most likely a develop version rather than a version from the dawn of time.

@becker33 will have to comment on why unresolved versions are allowed at all.