@scheibelp 2022-06-28
>> I'm concerned with the intent of this PR: when would it be valid to compare an unresolved git version?

> That has always been the default, except that it used to be essentially Version("''.0") or some similar version less than zero. 

@becker33 do you know why this is accounted for (i.e. why not just fail if we can't resolve the commit)? I didn't see an explanation of it in https://github.com/spack/spack/commit/ef5ad4eb34b351ba650ca92409c0295d5569efe4 (where the Git version commit functionality was originally added).